requirements.txt
fastapi==0.111.0
bcrypt==4.2.0
uvicorn==0.30.1
black==24.4.2
isort==5.13.2
SQLAlchemy==2.0.36
passlib==1.7.4
email_validator==2.2.0
cryptography==43.0.3
python-jose==3.3.0